Enhancement: Udpate auto connect to only permit one wallet connection for larpmode
